AP_Baro: fixed some signed/unsigned warnings
and a potential bug on timer wrap

+++ libraries/AP_Baro/AP_Baro_MS5611.h
@@ -27,7 +27,7 @@ class AP_Baro_MS5611 : public AP_Baro
   static bool _updated;
   static uint32_t _s_D1, _s_D2;
   static uint8_t _state;
-  static long _timer;
+  static uint32_t _timer;
   /* Gates access to asynchronous state: */
   static bool _sync_access;
